+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Relevant HTML title name.
 - Relevant HTML favicon.
+- State input is now a dropdown menu
 
 ## [0.3.1] - 07.13.2023